fix: modularize the code for writing unit tests
